% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisample.R
\name{align_counts_coords}
\alias{align_counts_coords}
\title{Align counts and coordinates}
\usage{
align_counts_coords(counts, coords_df)
}
\arguments{
\item{counts}{A sparse genes x spots matrix with column names as barcodes.}

\item{coords_df}{A data.frame with a column called barcode.}
}
\value{
A list with two elements:
\describe{
\item{counts}{Filtered counts matrix with columns restricted to common barcodes.}
\item{coords}{Filtered coordinate data.frame with rows in the same order as the counts columns.}
}
}
\description{
Align a sparse counts matrix and a coordinates data frame so that they use
the same set of barcodes in the same order. The coordinates are reordered
to match the column order of the counts matrix.
}
\keyword{internal}
