% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdutils.R
\name{calcAllThresholds}
\alias{calcAllThresholds}
\title{Compute the thresholds for all columns in a data frame}
\usage{
calcAllThresholds(df, minvals = 0.01, maxvals = 0.99, ...)
}
\arguments{
\item{df}{A data frame with pattern values (optionally
with x, y, barcode columns)}

\item{minval}{Minimum value for quantile threshold}

\item{maxval}{Maximum value for quantile threshold}
}
\value{
A list containing the computed thresholds for each pattern
}
\description{
This function computes the thresholds for all columns in a
data frame. The data frame could be an spPatterns object or an spInfluence
object.
}
