% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdutils.R
\name{calcIMscores.HD}
\alias{calcIMscores.HD}
\title{Calculate interaction scores for a specific pattern pair}
\usage{
calcIMscores.HD(
  data,
  patHotspots,
  infHotspots,
  patternpair,
  avoid_confounders = FALSE,
  ...
)
}
\arguments{
\item{data}{A numeric matrix with genes as rows and barcodes as columns.}

\item{patHotspots}{A data frame with pattern hotspots, containing columns for x, y, and barcode.}

\item{infHotspots}{A data frame with influence hotspots, containing columns for x, y, and barcode.}

\item{patternpair}{A character vector of length 2 specifying the pattern pair to analyze.}

\item{avoid_confounders}{Logical (default=FALSE) indicating whether to avoid confounding effects due to colocalization.}

\item{...}{Additional parameters to pass to lower level functions.}
}
\value{
A data frame with interaction scores for the specified pattern pair.
}
\description{
This function calculates interaction scores for a specific pattern pair
using the \code{classifySpots} function to determine the region of each spot.
}
