% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdutils.R
\name{calcInfluence}
\alias{calcInfluence}
\title{Compute the spatial influence of a spatial feature}
\usage{
calcInfluence(spPatterns, optParams, ...)
}
\arguments{
\item{spPatterns}{A data frame containing x, y coordinates and pattern name}

\item{patternName}{The name of the pattern to analyze}
}
\description{
This function computes the spatial influence of a specified pattern
}
