% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_gene_set_score}
\alias{calculate_gene_set_score}
\title{calculate_gene_set_score}
\usage{
calculate_gene_set_score(
  IMscores,
  gene_sets,
  weighted = TRUE,
  method = c("geometric_mean", "arithmetic_mean")
)
}
\arguments{
\item{IMscores}{A matrix of interaction scores}

\item{genes}{A list of gene sets, where each set is a vector of gene names}
}
\value{
A matrix of mean interaction scores for genes in each gene set, with
attributes for log p-value sums and number of genes for later fisher combination
}
\description{
Calculate the mean interaction score for a set of genes
}
