% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R, R/hdutils.R
\name{calculate_influence}
\alias{calculate_influence}
\alias{calculate_influence,SpaceMarkersExperiment-method}
\alias{calculate_influence,data.frame-method}
\title{Compute the spatial influence of a spatial feature}
\usage{
calculate_influence(spPatterns, optParams = NULL, ...)

\S4method{calculate_influence}{SpaceMarkersExperiment}(spPatterns, optParams = NULL, ...)

\S4method{calculate_influence}{data.frame}(spPatterns, optParams = NULL, ...)
}
\arguments{
\item{spPatterns}{A data frame containing x, y coordinates and pattern name}

\item{optParams}{A data frame with optimal parameters for the pattern}

\item{...}{Additional parameters for the Smooth function}
}
\value{
A data frame with the spatial influence of the specified pattern
}
\description{
This function computes the spatial influence of a specified pattern
}
