% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R, R/utils.R
\name{calculate_lr_scores}
\alias{calculate_lr_scores}
\alias{calculate_lr_scores,SpaceMarkersExperiment-method}
\alias{calculate_lr_scores,ANY-method}
\title{calculate_lr_scores}
\usage{
calculate_lr_scores(
  ligand_scores,
  receptor_scores = NULL,
  lr_pairs = NULL,
  ligand_test = NULL,
  method = "geometric_mean",
  weighted = TRUE
)

\S4method{calculate_lr_scores}{SpaceMarkersExperiment}(
  ligand_scores,
  receptor_scores = NULL,
  lr_pairs = NULL,
  ligand_test = NULL,
  method = "geometric_mean",
  weighted = TRUE
)

\S4method{calculate_lr_scores}{ANY}(
  ligand_scores,
  receptor_scores = NULL,
  lr_pairs = NULL,
  ligand_test = NULL,
  method = "geometric_mean",
  weighted = TRUE
)
}
\arguments{
\item{ligand_scores}{Output from getGeneSetScore for ligands}

\item{receptor_scores}{Output from getGeneSetScore for receptors}

\item{lr_pairs}{Data frame with columns 'ligand' and 'receptor'}

\item{ligand_test}{Character; specifies the type of test for ligand overexpression. Options are "greater" or "two.sided"}

\item{method}{Character; specifies the aggregation method for L-R scores. Options are "
geometric_mean" or "arithmetic_mean"}

\item{weighted}{Logical; if TRUE, L-R scores are weighted by their occurrence in
multiple L-R pairs}
}
\value{
Data frame with L-R scores and p-values
}
\description{
Calculate L-R pair scores using Fisher's method
}
\details{
This function computes L-R pair scores by combining ligand and receptor
overexpression scores using either geometric or arithmetic mean. It can also
weight L-R pairs based on their presence in multiple pairs to reduce bias from
promiscuous ligands or receptors.
}
\examples{
set.seed(1)
lr_pairs <- data.frame(
    ligand          = c("L1", "L2"),
    receptor        = c("R1", "R2"),
    ligand.symbol   = c("L1", "L2"),
    receptor.symbol = c("R1", "R2"),
    row.names       = c("L1_R1", "L2_R2"))
ls <- matrix(runif(4), 2, 2,
             dimnames = list(rownames(lr_pairs),
                             c("A_near_B", "B_near_A")))
rs <- matrix(runif(4), 2, 2,
             dimnames = list(rownames(lr_pairs), c("A", "B")))
calculate_lr_scores(ls, rs, lr_pairs = lr_pairs,
                    method = "arithmetic_mean", weighted = FALSE)
}
