% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdutils.R
\name{calculate_thresholds}
\alias{calculate_thresholds}
\title{Compute the thresholds for all columns in a data frame}
\usage{
calculate_thresholds(df, minvals = 0.01, maxvals = 0.99, ...)
}
\arguments{
\item{df}{A data frame with pattern values (optionally
with x, y, barcode columns)}

\item{minvals}{Minimum value for quantile threshold}

\item{maxvals}{Maximum value for quantile threshold}

\item{...}{Additional parameters to pass to lower level functions}
}
\value{
A list containing the computed thresholds for each pattern
}
\description{
This function computes the thresholds for all columns in a
data frame. The data frame could be an spPatterns object or an spInfluence
object.
}
\examples{
set.seed(1)
spPatterns <- data.frame(
    barcode = paste0("s", 1:50),
    x = runif(50), y = runif(50),
    Pattern_1 = rnorm(50), Pattern_2 = rnorm(50))
calculate_thresholds(spPatterns)
}
