% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cogaps_result}
\alias{cogaps_result}
\title{Latent Feature Space for each pattern}
\format{
CogapsResult object with 24228 features and 6 samples:
\describe{
\item{featureLoadings}{Data frame of Gene for each pattern}
\item{sampleFactors}{Data frame of cell barcodes and the 5 patterns}
}
}
\value{
A matrix of statistics for each pattern across each barcode
}
\description{
A CoGAPS object where the major requirements for SpaceMarkers are the
matrices of genes, barcodes and patterns learned of the latent-feature space
}
