% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{curated_genes}
\alias{curated_genes}
\title{Curated Genes for example purposes}
\format{
A vector with 114 pre-selected genes
}
\value{
a vector of genes
}
\description{
A vector with genes selected based on previous runs of SpaceMarkers on the
Visium 10x breast ductal carcinoma spatial transcriptomics dataset
}
