% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{DE_p1_p5_enrichment_df}
\alias{DE_p1_p5_enrichment_df}
\title{Interacting Genes including the KW-test genes}
\format{
A dataframe of 20 rows (the top 20 genes) and 18 columns:
\describe{
\item{Gene}{interacting genes}
\item{Pattern_1 x Pattern_5}{vsBoth,vsPattern_1,vsPattern5,FALSE}
\item{KW stats}{...}
\item{Dunns stats}{...}
\item{SpaceMarkersMetric}{|Dunn.zP1|x|Dunn.zP2|x2^-min(Dunn.zP1,Dunn.zP2)}
}
}
\value{
Dataframe summarizing latent feature interactions
}
\description{
A dataframe with:
Interactions between Pattern_1 and Pattern_5 from CoGAPS and
statistics about the interacting genes associated with them.
}
