% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R
\name{directed_scores}
\alias{directed_scores}
\alias{directed_scores<-}
\alias{directed_scores,SpaceMarkersExperiment-method}
\alias{directed_scores<-,SpaceMarkersExperiment-method}
\title{Access directed interaction scores on a SpaceMarkersExperiment}
\usage{
directed_scores(x, ...)

directed_scores(x) <- value

\S4method{directed_scores}{SpaceMarkersExperiment}(x)

\S4method{directed_scores}{SpaceMarkersExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{SpaceMarkersExperiment} object.}

\item{...}{Not used.}

\item{value}{A data.frame of directed interaction scores.}
}
\value{
For the getter, a data.frame of directed scores or \code{NULL}.
For the setter, the modified \code{SpaceMarkersExperiment}.

A data.frame of directed interaction scores, or NULL.
}
\description{
Read (\code{directed_scores(x)}) or write
(\code{directed_scores(x) <- value}) the data.frame of directed
interaction scores stored in \code{x@spacemarkers$results$directed_scores}.
}
