% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceMarkers.R
\name{.apply_sme_filters}
\alias{.apply_sme_filters}
\title{Filter an SME by gene expression / barcode match and (re)compute spatial params}
\usage{
.apply_sme_filters(
  sme,
  genes = NULL,
  min.gene.expr = 10,
  spatialDir = "spatial",
  pattern = "scalefactors_json.json",
  sigma = NULL,
  threshold = 4,
  resolution = "fullres"
)
}
\description{
Filter an SME by gene expression / barcode match and (re)compute spatial params
}
\keyword{internal}
