% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdutils.R
\name{.calc_threshold}
\alias{.calc_threshold}
\title{Compute the threshold for identifying outlier values or hotspots}
\usage{
.calc_threshold(df, minval = 0.01, maxval = 0.99, method = c("abs", "pct"))
}
\arguments{
\item{df}{A vector containing pattern values}

\item{minval}{Minimum value for quantile threshold}

\item{maxval}{Maximum value for quantile threshold}

\item{method}{Method to use for threshold calculation. Options are "abs" for absolute (default) and "pct" for percentile.}
}
\value{
A list containing the computed thresholds
}
\description{
This function computes the threshold for identifying outlier
values or hotspots by fitting a normal mixture model to the data.
}
