% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.get_BTME_features}
\alias{.get_BTME_features}
\title{.get_BTME_features
Load features BayesTME object}
\usage{
.get_BTME_features(hf)
}
\description{
.get_BTME_features
Load features BayesTME object
}
\keyword{internal}
