% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.get_cogaps_features}
\alias{.get_cogaps_features}
\title{.get_cogaps_features
Load features CoGAPS object}
\usage{
.get_cogaps_features(obj)
}
\value{
Numeric matrix of CoGAPS sample factors (rows = barcodes,
cols = patterns).
}
\description{
.get_cogaps_features
Load features CoGAPS object
}
\keyword{internal}
