% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.get_csv_features}
\alias{.get_csv_features}
\title{.get_csv_features
Load features from dataframe}
\usage{
.get_csv_features(obj)
}
\value{
data.frame of numeric features with barcodes as rownames.
}
\description{
.get_csv_features
Load features from dataframe
}
\keyword{internal}
