% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.get_seurat_features}
\alias{.get_seurat_features}
\title{.get_seurat_features
Load features Seurat object}
\usage{
.get_seurat_features(obj)
}
\value{
data.frame of \code{_Feature}-suffixed metadata columns extracted
from \code{slot(obj, "meta.data")}.
}
\description{
.get_seurat_features
Load features Seurat object
}
\keyword{internal}
