% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.getBTMEfeatures}
\alias{.getBTMEfeatures}
\title{.getBTMEfeatures
Load features BayesTME object}
\usage{
.getBTMEfeatures(hf)
}
\description{
.getBTMEfeatures
Load features BayesTME object
}
\keyword{internal}
