% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.getCogapsFeatures}
\alias{.getCogapsFeatures}
\title{.getCogapsFeatures
Load features CoGAPS object}
\usage{
.getCogapsFeatures(obj)
}
\description{
.getCogapsFeatures
Load features CoGAPS object
}
\keyword{internal}
