% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.getCSVfeatures}
\alias{.getCSVfeatures}
\title{.getDfFeatures
Load features from dataframe}
\usage{
.getCSVfeatures(obj)
}
\description{
.getDfFeatures
Load features from dataframe
}
\keyword{internal}
