% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.getSeuratFeatures}
\alias{.getSeuratFeatures}
\title{.getSeuratFeatures
Load features Seurat object}
\usage{
.getSeuratFeatures(obj)
}
\description{
.getSeuratFeatures
Load features Seurat object
}
\keyword{internal}
