% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.infer_method}
\alias{.infer_method}
\title{inferMethod
Infer the method used to obtain spatial features}
\usage{
.infer_method(spObject, method)
}
\value{
Character of length one: one of \code{"SpatialExperiment"},
\code{"BayesTME"}, \code{"CoGAPS"}, \code{"Seurat"}, or \code{"CSV"}.
}
\description{
inferMethod
Infer the method used to obtain spatial features
}
\keyword{internal}
