% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.inferMethod}
\alias{.inferMethod}
\title{inferMethod
Infer the method used to obtain spatial features}
\usage{
.inferMethod(spObject, method)
}
\description{
inferMethod
Infer the method used to obtain spatial features
}
\keyword{internal}
