% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSpaceMarkers.R
\name{.process_sme_input}
\alias{.process_sme_input}
\title{Process input from a SpaceMarkersExperiment}
\usage{
.process_sme_input(
  sme,
  genes = NULL,
  min.gene.expr = 10,
  spatialDir = "spatial",
  pattern = "scalefactors_json.json",
  sigma = NULL,
  threshold = 4,
  resolution = "fullres"
)
}
\description{
Process input from a SpaceMarkersExperiment
}
\keyword{internal}
