% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceMarkersExperiment-methods.R
\name{.sme_get_expr}
\alias{.sme_get_expr}
\title{Extract expression matrix from a SpaceMarkersExperiment}
\usage{
.sme_get_expr(sme, assay_name = "logcounts")
}
\arguments{
\item{sme}{A \code{SpaceMarkersExperiment} object.}

\item{assay_name}{Character name of the assay to extract. Default "logcounts".}
}
\value{
A matrix or sparse matrix of expression values.
}
\description{
Extract expression matrix from a SpaceMarkersExperiment
}
\keyword{internal}
