% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceMarkersExperiment-methods.R
\name{.sme_to_spPatterns}
\alias{.sme_to_spPatterns}
\title{Reconstruct legacy spPatterns data.frame from a SpaceMarkersExperiment}
\usage{
.sme_to_spPatterns(sme)
}
\arguments{
\item{sme}{A \code{SpaceMarkersExperiment} object.}
}
\value{
A data.frame with columns barcode, y, x, and pattern columns.
}
\description{
Reconstruct legacy spPatterns data.frame from a SpaceMarkersExperiment
}
\keyword{internal}
