% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSpaceMarkers.R
\name{.undirected_SpaceMarkers_sme}
\alias{.undirected_SpaceMarkers_sme}
\title{Undirected workflow on a SpaceMarkersExperiment}
\usage{
.undirected_SpaceMarkers_sme(
  sme,
  cpus = 1,
  genes = NULL,
  min.gene.expr = 10,
  spatialDir = "spatial",
  pattern = "scalefactors_json.json",
  sigma = NULL,
  threshold = 4,
  resolution = "fullres",
  returnSME = TRUE,
  ...
)
}
\description{
Undirected workflow on a SpaceMarkersExperiment
}
\keyword{internal}
