% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceMarkers.R
\name{.wrap_directed_result}
\alias{.wrap_directed_result}
\title{Wrap legacy directed result as SME (when file-path input used with returnSME=TRUE)}
\usage{
.wrap_directed_result(
  IMscores,
  features,
  data,
  resolution,
  version,
  h5filename
)
}
\description{
Wrap legacy directed result as SME (when file-path input used with returnSME=TRUE)
}
\keyword{internal}
