% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes_of_interest_nonparametric_fast.R
\name{find_genes_of_interest_nonparametric_fast}
\alias{find_genes_of_interest_nonparametric_fast}
\title{find_genes_of_interest_nonparametric_fast
Calculate ...}
\usage{
find_genes_of_interest_nonparametric_fast(testMat, goodGenes = NULL, region)
}
\arguments{
\item{testMat}{...}

\item{goodGenes}{...}

\item{region}{...}
}
\value{
a list of ...
}
\description{
This function calculates ...
}
