% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R, R/hdutils.R
\name{find_hotspots_gmm}
\alias{find_hotspots_gmm}
\alias{find_hotspots_gmm,SpaceMarkersExperiment-method}
\alias{find_hotspots_gmm,data.frame-method}
\title{Find hotspots for all patterns or influences based on values}
\usage{
find_hotspots_gmm(df, threshold = 0.1, ...)

\S4method{find_hotspots_gmm}{SpaceMarkersExperiment}(
  df,
  threshold = 0.1,
  ...,
  type = c("pattern", "influence"),
  minvals = NULL,
  maxvals = NULL
)

\S4method{find_hotspots_gmm}{data.frame}(df, threshold = 0.1, ...)
}
\arguments{
\item{df}{A data frame with pattern values (optionally
with x, y, barcode columns)}

\item{threshold}{a scalar or vector of thresholds for each column in the data frame.
Either user provided or the output of @calculate_thresholds}

\item{...}{Additional parameters to pass to lower level functions}

\item{minvals}{Minimum value for quantile threshold}

\item{maxvals}{Maximum value for quantile threshold}
}
\value{
a data frame with the same dimensions as the input data frame.
}
\description{
Convenience function to find hotspots for all spatial patterns
or influence dataframes based on provided thresholds
}
