% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInteractingGenes.R
\name{findAllHotspots}
\alias{findAllHotspots}
\title{Find hotpots for all spatial patterns}
\usage{
findAllHotspots(
  spPatterns,
  params = NULL,
  outlier = "positive",
  nullSamples = 100,
  includeSelf = TRUE,
  ...
)
}
\description{
Convenience function to find hotspots for all spatial patterns
}
