% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R, R/utils.R
\name{get_im_scores}
\alias{get_im_scores}
\alias{get_im_scores,SpaceMarkersExperiment-method}
\alias{get_im_scores,list-method}
\title{get_im_scores}
\usage{
get_im_scores(SpaceMarkers)

\S4method{get_im_scores}{SpaceMarkersExperiment}(SpaceMarkers)

\S4method{get_im_scores}{list}(SpaceMarkers)
}
\arguments{
\item{SpaceMarkers}{A list of SpaceMarkers objects}
}
\value{
A data frame with columns Gene and SpaceMarkersMetric
}
\description{
Get the interaction scores for SpaceMarkers
}
\examples{
example(get_pairwise_interacting_genes)
get_im_scores(SpaceMarkers)
}
