% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGeneSetScore}
\alias{getGeneSetScore}
\title{getGeneSetScore}
\usage{
getGeneSetScore(IMscores, genes)
}
\arguments{
\item{IMscores}{A matrix of interaction scores}

\item{genes}{A list of gene sets, where each set is a vector of gene names}
}
\value{
A vector of mean interaction scores for each gene set
}
\description{
Calculate the mean interaction score for a set of genes
}
