% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGseaPathways}
\alias{getGseaPathways}
\title{getGseaPathways}
\usage{
getGseaPathways(imscores, gene.sets, ...)
}
\arguments{
\item{imscores}{A data frame with genes as rownames and IMscores as columns}

\item{gene.sets}{a list where the names are pathway names and the values are
vectors of gene names}

\item{...}{params to pass to fgsea::fgsea}
}
\value{
A list of data frames with enriched pathways for each IMscore column
}
\description{
An fgsea wrapper to get enriched pathways for IMscores
}
