% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getIMScores}
\alias{getIMScores}
\title{getIMScores}
\usage{
getIMScores(SpaceMarkers)
}
\arguments{
\item{SpaceMarkers}{A list of SpaceMarkers objects}
}
\value{
A data frame with columns Gene and SpaceMarkersMetric
}
\description{
Get the interaction scores for SpaceMarkers
}
\examples{
example(getPairwiseInteractingGenes)
getIMScores(SpaceMarkers)

}
