% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdutils.R
\name{getOverlapScores.HD}
\alias{getOverlapScores.HD}
\title{getOverlapScores.HD}
\usage{
getOverlapScores.HD(
  patHotspots,
  infHotspots,
  patternList = NULL,
  method = c("relative-abundance", "differential-abundance", "absolute")
)
}
\arguments{
\item{patHotspots}{A data frame with columns x, y, barcode and pattern names}

\item{infHotspots}{A data frame with columns x, y, barcode and pattern names}

\item{patternList}{A character vector of pattern names to calculate overlap
scores for. If NULL, all patterns in patHotspots and infHotspots will be used.}

\item{method}{The method to calculate overlapping abundance scores. Options are
"relative-abundance", "differential-abundance" and "absolute"}
}
\value{
A data frame with columns pattern, influence and overlapping abundance
}
\description{
Calculate the overlap scores between patterns in hotspots
}
\details{
The function calculates the overlap scores between patterns hotspots
using the specified method. The default method is "relative-abundance"
}
\examples{
hotspots <- data.frame(x = c(1,2,3,4,5),
                        y = c(1,2,3,4,5),
                        barcode = c("A","B","C","D","E"),
                        pattern1 = c(1,0,1,0,1),
                        pattern2 = c(1,1,0,0,1))
getOverlapScores(hotspots)   
getOverlapScores(hotspots, c("pattern1","pattern2"))
}
