% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{getSpatialFeatures}
\alias{getSpatialFeatures}
\title{getSpatialFeatures
Load spatial features}
\usage{
getSpatialFeatures(filePath, method = "CoGAPS", featureNames = NULL)
}
\arguments{
\item{filePath}{A string path to the location of the file containing the
spatial features.}

\item{method}{A string specifying the method used to obtain spatial
features. e.g., "CoGAPS", "Seurat", or "BayesTME".}

\item{featureNames}{An array of strings specifying the column names
corresponding to the feature names. If input is NULL, in the case of CoGAPS
and BayesTME, all features are selected In the case of Seurat, all metadata
columns with "_Feature" suffix are selected.}
}
\value{
a matrix of spatial features with barcodes associated
with individual coordinates
}
\description{
This function loads spatial features from a file containing spatial features
}
\examples{
library(SpaceMarkers)
#CoGAPS data filePath
filePath <- system.file("extdata","CoGAPS_result.rds", 
package = "SpaceMarkers",mustWork = TRUE)
spFeatures <- getSpatialFeatures(filePath, method = "CoGAPS")
head(spFeatures)

}
