% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialParameters.R
\name{getSpatialParametersExternal}
\alias{getSpatialParametersExternal}
\title{getSpatialParametersExternal
Manually obtain Optimal Parameters for Interacting cells}
\usage{
getSpatialParametersExternal(
  spatialPatterns,
  visiumDir = ".",
  spatialDir = "spatial",
  pattern = "scalefactors_json.json",
  spotDiameter = NULL,
  threshold = 3
)
}
\arguments{
\item{spatialPatterns}{A data frame that contains the spatial coordinates
for each cell type. The column names must include 'x' and 'y' as well as a
set of numbered columns named  'Pattern_1.....N'.}

\item{visiumDir}{A string path specifying the location of the 10xVisium
directory}

\item{pattern}{A string specifying the name of the .json file}

\item{spotDiameter}{A numeric value specifying your desired sigma}

\item{threshold}{A numeric value specifying your hotspot threshold}

\item{spatial}{A string path specifying the location of the spatial folder
containing the .json file of the spot characteristics}
}
\value{
a numeric matrix of sigmaOpts - the optimal width of the gaussian
distribution, and the thresOpt - outlier threshold around the set of spots
for each pattern
}
\description{
This function calculates obtains a specified width of a distribution
(sigmaOpt) as well as the outlier threshold around the set of spots
(thresOpt) for each pattern from a latent feature space.
}
\examples{
library(SpaceMarkers)
# Create test data
cells <- c()
test_num <- 500
for(i in 1:test_num){
    cells[length(cells)+1] <- paste0("cell_",i)
}
spPatterns <- data.frame(barcode = cells,
y = runif(test_num, min=0, max=test_num),
x = runif(test_num, min=0, max=test_num),
Pattern_1 = runif(test_num, min=0, max=1),
Pattern_2 = runif(test_num, min=0, max=1) )
# Call the getSpatialParameters function with the test data
optParams <- getSpatialParametersExternal(spPatterns, spotDiameter = 10)

}
