% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load10XExpr}
\alias{load10XExpr}
\title{load10XExpr
load 10X Visium Expression Data}
\usage{
load10XExpr(visiumDir = NULL, h5filename = "filtered_feature_bc_matrix.h5")
}
\arguments{
\item{visiumDir}{A string path to the h5 file with expression information.}

\item{h5filename}{A string of the name of the h5 file in the directory.}
}
\description{
This log-transformed 10X Visium expression data from standard 10X Visium folder.
}
