% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R
\name{lr_scores}
\alias{lr_scores}
\alias{lr_scores<-}
\alias{lr_scores,SpaceMarkersExperiment-method}
\alias{lr_scores<-,SpaceMarkersExperiment-method}
\title{Access ligand-receptor scores on a SpaceMarkersExperiment}
\usage{
lr_scores(x, ...)

lr_scores(x) <- value

\S4method{lr_scores}{SpaceMarkersExperiment}(x)

\S4method{lr_scores}{SpaceMarkersExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{SpaceMarkersExperiment} object.}

\item{...}{Not used.}

\item{value}{A matrix of ligand-receptor pair scores.}
}
\value{
For the getter, a matrix of LR pair scores or \code{NULL}. For the
setter, the modified \code{SpaceMarkersExperiment}.

A matrix of ligand-receptor pair scores, or NULL.
}
\description{
Read (\code{lr_scores(x)}) or write (\code{lr_scores(x) <- value}) the
matrix of ligand-receptor pair scores stored in
\code{x@spacemarkers$results$lr_scores}.
}
