% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lrlist}
\alias{lrlist}
\title{Curated Ligand-receptor interaction genes
A list of vectors with genes associated with ligand-receptor interactions from CellChat database}
\format{
A list of length 3234 with each element being a vector of genes
}
\value{
A list of vectors of LR genes
}
\description{
Curated Ligand-receptor interaction genes
A list of vectors with genes associated with ligand-receptor interactions from CellChat database
}
