% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optParams_breast_cancer.R
\docType{data}
\name{optParams_breast_cancer}
\alias{optParams_breast_cancer}
\title{Optimal paramters of 5 patterns from CoGAPS.}
\format{
A data frame with 2 rows and 5 columns:
\describe{
\item{Pattern_1}{immune cell pattern paramters }
.
.
.
\item{Pattern_5}{invasive carcinoma lesion pattern paramters }

}
}
\usage{
optParams_breast_cancer
}
\description{
A dataset with the optimal width of the gaussian distribution (sigmaOpt) and
the outlier threshold around the set of spots (thresOpt) for each pattern
obtained from CoGAPS. CoGAPS was ran on spatial transcriptomic data from a
breast cancer sample.
}
\keyword{datasets}
