% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R
\name{overlap_scores}
\alias{overlap_scores}
\alias{overlap_scores<-}
\alias{overlap_scores,SpaceMarkersExperiment-method}
\alias{overlap_scores<-,SpaceMarkersExperiment-method}
\title{Access pattern overlap scores on a SpaceMarkersExperiment}
\usage{
overlap_scores(x, ...)

overlap_scores(x) <- value

\S4method{overlap_scores}{SpaceMarkersExperiment}(x)

\S4method{overlap_scores}{SpaceMarkersExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{SpaceMarkersExperiment} object.}

\item{...}{Not used.}

\item{value}{A data.frame of pattern overlap scores.}
}
\value{
For the getter, a data.frame of pattern overlap scores or
\code{NULL}. For the setter, the modified \code{SpaceMarkersExperiment}.

A data.frame of pattern overlap scores, or NULL.
}
\description{
Read (\code{overlap_scores(x)}) or write
(\code{overlap_scores(x) <- value}) the data.frame of pattern overlap
scores stored in \code{x@spacemarkers$results$overlap_scores}.
}
\examples{
sme <- SpaceMarkersExperiment(
    assays = list(logcounts = matrix(rpois(200, 2), 10, 20,
        dimnames = list(paste0("G", 1:10), paste0("s", 1:20)))),
    spatialCoords = matrix(runif(40), 20, 2,
        dimnames = list(NULL, c("y", "x"))))
overlap_scores(sme) <- data.frame(pattern1 = "Pattern_1",
    pattern2 = "Pattern_2", overlapScore = 0.5)
overlap_scores(sme)
}
