% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pickImage}
\alias{pickImage}
\title{pickImage}
\usage{
pickImage(sp_dir, res)
}
\arguments{
\item{sp_dir}{path to the spatial directory}

\item{res}{a character string specifying the resolution of the image}
}
\value{
a character string of the image file name
}
\description{
The function picks the appropriate histology image file from the
spatial directory based on the specified resolution.
}
