% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotIMScores}
\alias{plotIMScores}
\title{plotIMScores}
\usage{
plotIMScores(
  df,
  interaction,
  cutOff = 0,
  nGenes = 20,
  geneText = 12,
  metricText = 12,
  increments = 1,
  out = NULL
)
}
\arguments{
\item{df}{A data frame with columns Gene and SpaceMarkersMetric}

\item{interaction}{The interaction to plot}

\item{cutOff}{The cut off value for the plot}

\item{nGenes}{The number of genes to plot}

\item{geneText}{The font size for the gene text}

\item{metricText}{The font size for the metric text}

\item{increments}{The increments for the y-axis}

\item{out}{The output path for the plot}
}
\description{
Plot the top SpaceMarkers IMScores
}
\examples{
example(getPairwiseInteractingGenes)
plotIMScores(getIMScores(SpaceMarkers), "PDAC", 0.5, 20, 12, 12, 1)

}
