% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotOverlapScores}
\alias{plotOverlapScores}
\title{plotOverlapScores}
\usage{
plotOverlapScores(
  df,
  title = "Spatial Overlap Scores",
  out = NULL,
  fontsize = 15,
  tiletext = 6
)
}
\arguments{
\item{df}{A data frame with columns pattern1, pattern2 and overlapScore}

\item{title}{The title of the plot}

\item{out}{The output path for the plot}

\item{fontsize}{The font size of the plot}

\item{tiletext}{The font size within the tile}
}
\value{
A ggplot object
}
\description{
Plot the overlap scores between patterns in hotspots
}
\examples{
df <- data.frame(pattern1 = c("pattern1","pattern1","pattern2","pattern2"), 
                 pattern2 = c("pattern1","pattern2","pattern1","pattern2"),
                 overlapScore = c(0.5,0.7,0.3,0.9))
plotOverlapScores(df)
plotOverlapScores(df, "Overlap Scores", "overlapScores.png", 15)
}
