% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotScoresHeatmap}
\alias{plotScoresHeatmap}
\title{plotScoresHeatmap}
\usage{
plotScoresHeatmap(
  df,
  interactions = NULL,
  geneCol = NULL,
  geneCutOff = 20,
  out = "ScoresHeatmap.png",
  fontsize_row = 7,
  fontsize_col = 7,
  cluster_cols = TRUE,
  cluster_rows = TRUE,
  ...
)
}
\arguments{
\item{df}{A data frame with genes as rownames and IMscores as columns}

\item{interactions}{a vector of column names to subset the heatmap}

\item{geneCol}{a character specifying the column with the gene names.
Default: NULL and assumes rownames}

\item{geneCutOff}{Number of genes to plot based on max score, Default: 20}

\item{out}{Filename of heatmap, Default: 'ScoresHeatmap.png'}

\item{fontsize_row}{Size of row text, Default: 7}

\item{fontsize_col}{Size of column text N, Default: 7}

\item{cluster_cols}{Whether or not to cluster columns on the heatmap,
Default: TRUE}

\item{cluster_rows}{Whether or not to clutser rows on the heatmap,
Default: TRUE}

\item{...}{Additional parameters to pass to pheatmap::pheatmap}
}
\value{
A Pheatmap object
}
\description{
Generates top IMscores heatmap based on the maximum score per
row where the number of rows is defined by geneCutOff
}
\examples{
example(getPairwiseInteractingGenes)
imScores <- getIMScores(SpaceMarkers)
plotScoresHeatmap(imScores, geneCutOff = 15, out = "IMscoresHeatmap.png")
}
