% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceMarkersExperiment-methods.R
\name{reexports}
\alias{reexports}
\alias{assay}
\alias{assays}
\alias{colData}
\alias{rowData}
\alias{spatialCoords}
\alias{spatialCoords<-}
\alias{imgRaster}
\title{SummarizedExperiment / SpatialExperiment accessors re-exported}
\description{
These generics are defined in \pkg{SummarizedExperiment} and
\pkg{SpatialExperiment} and re-exported here so that
\code{library(SpaceMarkers)} alone is enough to call them on a
\code{SpaceMarkersExperiment} without also attaching their source packages.
}
\keyword{internal}
