% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{residual_p1_p5_enrichment_genesOnly}
\alias{residual_p1_p5_enrichment_genesOnly}
\title{Interacting Genes}
\format{
A vector of length 21331
}
\value{
A vector of genes returned from enrichment analysis
}
\description{
A vector with:
Genes identified between Pattern_1 and Pattern_5 genes from
getInteractingGenes(...,analysis = "enrichment")
}
