% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceMarkers_DEmode_overlap.R
\name{SpaceMarkers_DEmode_overlap}
\alias{SpaceMarkers_DEmode_overlap}
\title{Interacting Genes for each Pattern.}
\format{
A list of a list of 4 data frames and 1 matrix:
\describe{
\item{interacting genes list of 4 dfs}{genes,CoGAPS patterns and stats}
\item{hotspot Regions matrix}{CoGAPS hotspots for each spot}
}
}
\value{
List of data frames summarizing latent feature interactions
}
\description{
A list with:
\enumerate{
\item A list of 4 each containing a data frame of Pattern interactions from
CoGAPS and statistics about the interacting genes associated with them.
}
}
\details{
\enumerate{
\item A matrix of the hotspot regions from CoGAPS for each spot
}
}
