% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R
\name{undirected_scores}
\alias{undirected_scores}
\alias{undirected_scores<-}
\alias{undirected_scores,SpaceMarkersExperiment-method}
\alias{undirected_scores<-,SpaceMarkersExperiment-method}
\title{Access undirected interaction scores on a SpaceMarkersExperiment}
\usage{
undirected_scores(x, ...)

undirected_scores(x) <- value

\S4method{undirected_scores}{SpaceMarkersExperiment}(x)

\S4method{undirected_scores}{SpaceMarkersExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{SpaceMarkersExperiment} object.}

\item{...}{Not used.}

\item{value}{A data.frame of undirected interaction scores.}
}
\value{
For the getter, a data.frame of undirected scores or \code{NULL}.
For the setter, the modified \code{SpaceMarkersExperiment}.

A data.frame of undirected interaction scores (genes x pattern
pairs), or NULL.
}
\description{
Read (\code{undirected_scores(x)}) or write
(\code{undirected_scores(x) <- value}) the data.frame of undirected
interaction scores stored in \code{x@spacemarkers$results$undirected_scores}.
}
\examples{
sme <- SpaceMarkersExperiment(
    assays = list(logcounts = matrix(rpois(200, 2), 10, 20,
        dimnames = list(paste0("G", 1:10), paste0("s", 1:20)))),
    spatialCoords = matrix(runif(40), 20, 2,
        dimnames = list(NULL, c("y", "x"))))
undirected_scores(sme) <- data.frame(
    Gene = paste0("G", 1:5), Pattern_1_Pattern_2 = runif(5))
head(undirected_scores(sme))
}
