% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{as-SpatialExperiment-SpaceMarkersExperiment}
\alias{as-SpatialExperiment-SpaceMarkersExperiment}
\title{Coercion from SpatialExperiment to SpaceMarkersExperiment}
\description{
Coercion from SpatialExperiment to SpaceMarkersExperiment
}
