% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_overlap_scores}
\alias{calc_overlap_scores}
\title{calc_overlap_scores}
\usage{
calc_overlap_scores(
  hotspots,
  patternList = NULL,
  method = c("Szymkiewicz-Simpson", "Jaccard", "Sorensen-Dice", "Ochiai", "absolute")
)
}
\arguments{
\item{hotspots}{A data frame with columns x, y, barcode and pattern names}

\item{patternList}{A character vector of pattern names to calculate overlap
scores for}

\item{method}{The method to calculate overlap scores. Options are
"Szymkiewicz-Simpson", "Jaccard", "Sorensen-Dice", "Ochiai" and "absolute"}
}
\value{
A matrix of overlap scores (patterns × patterns)
}
\description{
Calculate the overlap scores between patterns in hotspots
}
\details{
The function calculates the overlap scores between patterns hotspots
using the specified method. The default method is "Szymkiewicz-Simpson"
overlap coefficient.
}
\examples{
hotspots <- data.frame(x = c(1,2,3,4,5),
                        y = c(1,2,3,4,5),
                        barcode = c("A","B","C","D","E"),
                        pattern1 = c(1,0,1,0,1),
                        pattern2 = c(1,1,0,0,1))
calc_overlap_scores(hotspots = hotspots)   
}
