% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdutils.R
\name{calcAllIMscores.HD}
\alias{calcAllIMscores.HD}
\title{Calculate interaction scores for all pattern pairs}
\usage{
calcAllIMscores.HD(data, patHotspots, infHotspots, patternPairs = NULL)
}
\arguments{
\item{data}{A numeric matrix with genes as rows and barcodes as columns.}

\item{patHotspots}{A data frame with pattern hotspots, containing columns for x, y, and barcode.}

\item{infHotspots}{A data frame with influence hotspots, containing columns for x, y, and barcode.}

\item{patternPairs}{A data frame with pattern pairs to calculate interaction scores for.}
}
\value{
A data frame with interaction scores for all pattern pairs.
}
\description{
This function calculates interaction scores for all pattern pairs
using the \code{calcIMscores.HD} function. It can run in parallel if BiocParallel is available.
}
