% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdutils.R
\name{calcInfluence}
\alias{calcInfluence}
\title{Compute the spatial influence of a spatial feature}
\usage{
calcInfluence(spPatterns, optParams, ...)
}
\arguments{
\item{spPatterns}{A data frame containing x, y coordinates and pattern name}

\item{optParams}{A data frame with optimal parameters for the pattern}

\item{...}{Additional parameters for the Smooth function}
}
\value{
A data frame with the spatial influence of the specified pattern
}
\description{
This function computes the spatial influence of a specified pattern
}
