% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdutils.R
\name{calcThresholds}
\alias{calcThresholds}
\title{Compute the threshold for identifying outlier values or hotspots}
\usage{
calcThresholds(df, minval = 0.01, maxval = 0.99)
}
\arguments{
\item{df}{A data frame containing x, y coordinates and pattern name}

\item{minval}{Minimum value for quantile threshold}

\item{maxval}{Maximum value for quantile threshold}

\item{pat}{The name of the pattern to analyze}
}
\value{
A list containing the computed thresholds
}
\description{
This function computes the threshold for identifying outlier
values or hotspots by fitting a normal mixture model to the data.
}
