% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_lr_scores}
\alias{calculate_lr_scores}
\title{calculate_lr_scores}
\usage{
calculate_lr_scores(
  ligand_scores,
  receptor_scores,
  lr_pairs,
  ligand_test = c("greater", "two.sided"),
  method = c("geometric_mean", "arithmetic_mean"),
  weighted = TRUE
)
}
\arguments{
\item{lr_pairs}{Data frame with columns 'ligand' and 'receptor'}

\item{ligand_results}{Output from getGeneSetScore for ligands}

\item{receptor_results}{Output from getGeneSetScore for receptors}
}
\value{
Data frame with L-R scores and p-values
}
\description{
Calculate L-R pair scores using Fisher's method
}
