% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{.get_seurat_features}
\alias{.get_seurat_features}
\title{.get_seurat_features
Load features Seurat object}
\usage{
.get_seurat_features(obj)
}
\description{
.get_seurat_features
Load features Seurat object
}
\keyword{internal}
