% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSpaceMarkers.R
\name{.wrap_undirected_result}
\alias{.wrap_undirected_result}
\title{Wrap legacy undirected result as SME (when file-path input used with returnSME=TRUE)}
\usage{
.wrap_undirected_result(
  IMScores,
  features,
  data,
  resolution,
  version,
  h5filename
)
}
\description{
Wrap legacy undirected result as SME (when file-path input used with returnSME=TRUE)
}
\keyword{internal}
