% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_genes_of_interest_nonparametric_fast.R
\name{find_genes_of_interest_nonparametric_fast}
\alias{find_genes_of_interest_nonparametric_fast}
\title{find_genes_of_interest_nonparametric_fast
Identify genes associated with pattern interaction.}
\usage{
find_genes_of_interest_nonparametric_fast(
  testMat,
  goodGenes = NULL,
  region,
  fdr.level = 0.05
)
}
\arguments{
\item{testMat}{...}

\item{goodGenes}{...}

\item{region}{...}

\item{fdr.level}{False Discovery Rate. The default value is 0.05.}
}
\value{
a list of genes exhibiting significantly higher values of testMat in the Interaction region of the two #' patterns compared to regions with exclusive influence from either pattern.
}
\description{
This function identifies genes exhibiting significantly higher values of testMat in the Interaction region of the two
patterns compared #' to regions with exclusive influence from either pattern. It uses Kruskal-Wallis test followed by
posthoc analysis using Dunn's Test to identify the genes.
}
