% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInteractingGenes.R
\name{findAllHotspots}
\alias{findAllHotspots}
\title{Find hotpots for all spatial patterns}
\usage{
findAllHotspots(
  spPatterns,
  params = NULL,
  outlier = "positive",
  nullSamples = 100,
  includeSelf = TRUE,
  ...
)
}
\arguments{
\item{spPatterns}{A data frame that contains the spatial coordinates
and metrics for spatial features (cell types/cell processes). The column
names must include 'x' and 'y' as well as the spatially varying features.}

\item{params}{a matrix with dimensions 2 X N, where N is the number of
spatial patterns with optimal parameters. The first row contains the kernel
width 'sigmaOpt' for each pattern, and the second row is the threshOpt
(outlier threshold) for each pattern. Users can also input their preferred
param values. The default value is NULL.}

\item{outlier}{a character string specifying whether to apply the outlier
threshold to the kernel density distribution in a one-sided manner (specify
'positive' the default) or in a two sided manner (specify 'two.sided').}

\item{nullSamples}{a numeric values specifying the number of spatial patterns
to randomly sample for a null distribution.}

\item{includeSelf}{a logic value specifying whether to consider the spatial
influence the pattern has on surrounding regions only (set to FALSE), or
whether to also consider the influence of the pattern itself (set to TRUE,
the default).}

\item{...}{Arguments passed to methods}
}
\description{
Convenience function to find hotspots for all spatial patterns
}
