% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R, R/utils.R
\name{get_im_scores}
\alias{get_im_scores}
\alias{get_im_scores,SpaceMarkersExperiment-method}
\alias{get_im_scores,list-method}
\title{get_im_scores}
\usage{
get_im_scores(SpaceMarkers)

\S4method{get_im_scores}{SpaceMarkersExperiment}(SpaceMarkers)

\S4method{get_im_scores}{list}(SpaceMarkers)
}
\arguments{
\item{SpaceMarkers}{A list returned by
\code{\link{get_pairwise_interacting_genes}} (or a
\code{SpaceMarkersExperiment} with interactions populated).}
}
\value{
A data frame with columns Gene and SpaceMarkersMetric
}
\description{
Get the interaction scores for SpaceMarkers
}
\examples{
# Minimal mock of the list shape get_pairwise_interacting_genes returns:
sm_result <- list(
    list(patterns = c("A","B"),
         interacting_genes = list(data.frame(
             Gene = paste0("G", 1:5),
             SpaceMarkersMetric = sort(runif(5), decreasing = TRUE)))),
    list(patterns = c("A","C"),
         interacting_genes = list(data.frame(
             Gene = paste0("G", 1:5),
             SpaceMarkersMetric = sort(runif(5), decreasing = TRUE)))))
get_im_scores(sm_result)
}
