% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_im_scores}
\alias{get_im_scores}
\title{get_im_scores}
\usage{
get_im_scores(SpaceMarkers)
}
\arguments{
\item{SpaceMarkers}{A list of SpaceMarkers objects}
}
\value{
A data frame with columns Gene and SpaceMarkersMetric
}
\description{
Get the interaction scores for SpaceMarkers
}
\examples{
example(get_pairwise_interacting_genes)
get_im_scores(SpaceMarkers)
}
