% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_overlap_scores}
\alias{get_overlap_scores}
\title{get_overlap_scores}
\usage{
get_overlap_scores(
  hotspots = NULL,
  in_hotspots = NULL,
  pat_hotspots = NULL,
  patternList = NULL,
  method = c("Szymkiewicz-Simpson", "Jaccard", "Sorensen-Dice", "Ochiai", "absolute")
)
}
\arguments{
\item{hotspots}{A data frame with columns x, y, barcode and pattern names}

\item{in_hotspots}{Hotspots from the influence of patterns, Default: NULL}

\item{pat_hotspots}{Hotspots from the pattern itself, Default: NULL}

\item{patternList}{A character vector of pattern names to calculate overlap
scores for}

\item{method}{The method to calculate overlap scores. Options are
"Szymkiewicz-Simpson", "Jaccard", "Sorensen-Dice", "Ochiai" and "absolute"}
}
\value{
A data frame with columns pattern1, pattern2 and overlapScore
}
\description{
Calculate the overlap scores between patterns in hotspots
}
\details{
The function calculates the overlap scores between patterns hotspots
using the specified method. The default method is "Szymkiewicz-Simpson"
overlap coefficient.
}
\examples{
in_hotspots <- data.frame(x = c(1,2,3,4,5),
                          y = c(1,2,3,4,5),
                          barcode = c("A","B","C","D","E"),
                          pattern1 = c(1,0,1,0,1),
                          pattern2 = c(1,1,0,0,1))
pat_hotspots <- data.frame(x = c(1,2,3,4,5),y = c(1,2,3,4,5),
                          barcode = c("A","B","C","D","E"),
                           pattern1 = c(0.3,1,0,0,1),
                           pattern3 = c(0,NA,1,0,NA))
get_overlap_scores(hotspots = NULL, in_hotspots = in_hotspots, 
                   pat_hotspots = pat_hotspots, method = "absolute")

}
