% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInteractingGenes.R
\name{getInteractingGenes}
\alias{getInteractingGenes}
\title{getInteractingGenes
Calculate ...}
\usage{
getInteractingGenes(
  data,
  reconstruction = NULL,
  spatialPatterns,
  refPattern = "Pattern_1",
  mode = c("residual", "DE"),
  minOverlap = 50,
  hotspotRegions = NULL
)
}
\arguments{
\item{data}{...}

\item{reconstruction}{...}

\item{spatialPatterns}{...}

\item{refPattern}{a character string that specifies the pattern whose "interaction" with every other pattern we want to study.
The default value is "Pattern_1".}

\item{mode}{...}

\item{minOverlap}{...}

\item{hotspotRegions}{...}
}
\value{
a list of ...
}
\description{
This function calculates ...
}
