% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInteractingGenes.R
\name{getInteractingGenes}
\alias{getInteractingGenes}
\title{getInteractingGenes
Calculate ...}
\usage{
getInteractingGenes(
  data,
  reconstruction = NULL,
  spatialPatterns,
  optParams = NULL,
  refPattern = "Pattern_1",
  mode = c("residual", "DE"),
  minOverlap = 50,
  hotspotRegions = NULL
)
}
\arguments{
\item{data}{original spatial data matrix}

\item{reconstruction}{reconstruction of the data matrix from latent spaces
Required for "residual" mode}

\item{spatialPatterns}{...}

\item{optParams}{a matrix with dimensions 2 X N, where N is the number of patterns with optimal parameters for outlier
detection calculated from function getSpatialParameters(). The first row contains the kernel width sigmaOpt for each
pattern, ahd the second threshold is the threshOpt (outlier threshold) for each pattern. Users can also input their
preferred param values.
The default value is NULL.}

\item{refPattern}{a character string that specifies the pattern whose "interaction" with every other pattern we want
to study. The default value is "Pattern_1".}

\item{mode}{SpaceMarkers mode of operation. Possible values are "residual" or "DE".}

\item{minOverlap}{...}

\item{hotspotRegions}{...}
}
\value{
a list of ...
}
\description{
This function calculates ...
}
