% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialParameters.R
\name{getSpatialParameters}
\alias{getSpatialParameters}
\title{getSpatialParameters
Calculate the Optimal Parameters for Interacting Cells}
\usage{
getSpatialParameters(spatialPatterns, ...)
}
\arguments{
\item{spatialPatterns}{A data frame that contains the spatial coordinates
for each cell type. The column names must include 'x' and 'y' as well as a
set of numbered columns named  'Pattern_1.....N'.}

\item{...}{Arguments passed to methods}
}
\value{
a numeric matrix of sigmaOpts - the optimal width of the gaussian
distribution, and the thresOpt - outlier threshold around the set of spots
for each pattern
}
\description{
This function calculates the optimal width of the gaussian distribution
(sigmaOpt) as well as the outlier threshold around the set of spots
(thresOpt) for each pattern from a latent feature space.
}
\examples{
# Create test data
cells <- c()
test_num <- 500
for(i in 1:test_num){
    cells[length(cells)+1] <- paste0("cell_",i)
}
spPatterns <- data.frame(barcode = cells,
y = runif(test_num, min=0, max=test_num),
x = runif(test_num, min=0, max=test_num),
Pattern_1 = runif(test_num, min=0, max=1),
Pattern_2 = runif(test_num, min=0, max=1) )
# Call the getSpatialParameters function with the test data
optParams <- getSpatialParameters(spPatterns)

}
