% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialParameters.R
\name{getSpatialParameters}
\alias{getSpatialParameters}
\title{getSpatialParameters
Calculate ...}
\usage{
getSpatialParameters(spatialPatterns)
}
\arguments{
\item{spatialPatterns}{...}
}
\description{
This function calculates ...
}
