% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R
\name{hotspots}
\alias{hotspots}
\alias{hotspots<-}
\alias{hotspots,SpaceMarkersExperiment-method}
\alias{hotspots<-,SpaceMarkersExperiment-method}
\title{Access hotspot assignments on a SpaceMarkersExperiment}
\usage{
hotspots(x, ...)

hotspots(x, type = "undirected") <- value

\S4method{hotspots}{SpaceMarkersExperiment}(x, type = c("undirected", "pattern", "influence"))

\S4method{hotspots}{SpaceMarkersExperiment}(x, type = c("undirected", "pattern", "influence")) <- value
}
\arguments{
\item{x}{A \code{SpaceMarkersExperiment} object.}

\item{...}{Not used.}

\item{type}{Character: one of "undirected", "pattern", or "influence".}

\item{value}{A data.frame of hotspot assignments.}
}
\value{
For the getter, a data.frame or \code{NULL}. For the setter, the
modified \code{SpaceMarkersExperiment}.

A data.frame of hotspot assignments, or NULL.
}
\description{
Read (\code{hotspots(x, type)}) or write
(\code{hotspots(x, type) <- value}) the per-type hotspot data.frame stored
in \code{metadata(x)$hotspots[[type]]}. \code{type} selects one of
\code{"undirected"}, \code{"pattern"}, or \code{"influence"}.
}
\examples{
sme <- SpaceMarkersExperiment(
    assays = list(logcounts = matrix(rpois(200, 2), 10, 20,
        dimnames = list(paste0("G", 1:10), paste0("s", 1:20)))),
    spatialCoords = matrix(runif(40), 20, 2,
        dimnames = list(NULL, c("y", "x"))))
colnames(sme) <- paste0("s", seq_len(20))
hs <- data.frame(barcode = colnames(sme), x = 1:20, y = 1:20,
                 Pattern_1 = c(rep("hot", 10), rep(NA, 10)))
hotspots(sme, type = "undirected") <- hs
head(hotspots(sme, "undirected"))
}
