% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load10XCoords}
\alias{load10XCoords}
\title{load10XCoords
Load 10x Visium Spatial Coordinates}
\usage{
load10XCoords(visiumDir, resolution = "lowres")
}
\arguments{
\item{visiumDir}{A string path to the location of the folder containing the
spatial coordinates. The folder in your visiumDir must be named 'spatial' and
must contain files 'scalefactors_json.json' and 'tissue_positions_list.csv.'}

\item{resolution}{A string specifying which values to look for in the .json
object. Can be either lowres or highres.}
}
\value{
a data frame of the spatial coordinates ( x and y) for each spot/cell
}
\description{
This function loads spatial coordinates for each cell from a 10X Visium
spatial folder.
}
\examples{
main_10xlink <- "https://cf.10xgenomics.com/samples/spatial-exp/1.3.0"
sp_folder <- "Visium_FFPE_Human_Breast_Cancer"
sp_file <- "Visium_FFPE_Human_Breast_Cancer_spatial.tar.gz"
sp_url <- paste(c(main_10xlink,sp_folder,sp_file),collapse = "/")
# Spatial Coordinates
download.file(sp_url, basename(sp_url))
untar(basename(sp_url))
spCoords <- load10XCoords(visiumDir = ".")
unlink("spatial", recursive = TRUE)
unlink("Visium_FFPE_Human_Breast_Cancer_spatial.tar.gz")

}
