% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load10XCoords}
\alias{load10XCoords}
\title{load10XCoords
Load ...}
\usage{
load10XCoords(visiumDir, resolution = "lowres")
}
\arguments{
\item{visiumDir}{...}

\item{resolution}{...}
}
\description{
This function loads ...
}
