% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load10XExpr}
\alias{load10XExpr}
\title{load10XExpr
load log-transformed 10X Visium expression data from standard 10X Visium folder}
\usage{
load10XExpr(visiumDir = NULL, h5filename = "filtered_feature_bc_matrix.h5")
}
\arguments{
\item{visiumDir}{...}

\item{h5filename}{...}
}
\description{
This function loads ...
}
