% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{optParams_breast_cancer}
\alias{optParams_breast_cancer}
\title{Optimal paramters of 5 patterns from CoGAPS.}
\format{
A data frame with 2 rows and 5 columns:
\describe{
\item{Pattern_1}{immune cell pattern paramters }
\item{Pattern_2}{Disp.1 parameters}
\item{Pattern_3}{intraductal carcinoma (DCIS) parameters }
\item{Pattern_2}{Disp.2 parameters }
\item{Pattern_5}{invasive carcinoma lesion pattern paramters }

}
}
\value{
A matrix of optimal parameters for patterns identified by CoGAPS
}
\description{
A dataset with the optimal width of the gaussian distribution (sigmaOpt) and
the outlier threshold around the set of spots (thresOpt) for each pattern
obtained from CoGAPS. CoGAPS was ran on spatial transcriptomic data from a
breast cancer sample.
}
