% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R
\name{overlap_scores}
\alias{overlap_scores}
\alias{overlap_scores<-}
\alias{overlap_scores,SpaceMarkersExperiment-method}
\alias{overlap_scores<-,SpaceMarkersExperiment-method}
\title{Access pattern overlap scores on a SpaceMarkersExperiment}
\usage{
overlap_scores(x, ...)

overlap_scores(x) <- value

\S4method{overlap_scores}{SpaceMarkersExperiment}(x)

\S4method{overlap_scores}{SpaceMarkersExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{SpaceMarkersExperiment} object.}

\item{...}{Not used.}

\item{value}{A data.frame of pattern overlap scores.}
}
\value{
For the getter, a data.frame of pattern overlap scores or
\code{NULL}. For the setter, the modified \code{SpaceMarkersExperiment}.

A data.frame of pattern overlap scores, or NULL.
}
\description{
Read (\code{overlap_scores(x)}) or write
(\code{overlap_scores(x) <- value}) the data.frame of pattern overlap
scores stored in \code{x@spacemarkers$results$overlap_scores}.
}
