% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R
\name{SpaceMarkersExperiment-accessors}
\alias{SpaceMarkersExperiment-accessors}
\alias{spatial_patterns<-}
\alias{spatial_patterns,SpaceMarkersExperiment-method}
\alias{spatial_patterns<-,SpaceMarkersExperiment-method}
\title{Accessor generics for SpaceMarkersExperiment}
\usage{
spatial_patterns(x, ...)

spatial_patterns(x) <- value

\S4method{spatial_patterns}{SpaceMarkersExperiment}(x)

\S4method{spatial_patterns}{SpaceMarkersExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{SpaceMarkersExperiment} object.}

\item{value}{A data.frame or DataFrame of spatial pattern values. Row names
must match \code{colnames(x)}.}
}
\value{
A \code{DataFrame} of spatial pattern values per spot, or NULL if
pattern names have not been set.
}
\description{
Generics for accessing SpaceMarkersExperiment results.
}
