% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/SpaceMarkersExperiment-methods.R
\name{undirected_scores}
\alias{undirected_scores}
\alias{undirected_scores<-}
\alias{undirected_scores,SpaceMarkersExperiment-method}
\alias{undirected_scores<-,SpaceMarkersExperiment-method}
\title{Access undirected interaction scores on a SpaceMarkersExperiment}
\usage{
undirected_scores(x, ...)

undirected_scores(x) <- value

\S4method{undirected_scores}{SpaceMarkersExperiment}(x)

\S4method{undirected_scores}{SpaceMarkersExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{SpaceMarkersExperiment} object.}

\item{...}{Not used.}

\item{value}{A data.frame of undirected interaction scores.}
}
\value{
For the getter, a data.frame of undirected scores or \code{NULL}.
For the setter, the modified \code{SpaceMarkersExperiment}.

A data.frame of undirected interaction scores (genes x pattern
pairs), or NULL.
}
\description{
Read (\code{undirected_scores(x)}) or write
(\code{undirected_scores(x) <- value}) the data.frame of undirected
interaction scores stored in \code{x@spacemarkers$results$undirected_scores}.
}
